fix(packaging): Set Rxjs version to follow Angular's
When working with a common repository and so importing a some component of other application written in Angular, there is conflict with the Rxjs dependency version with Typescript.
Typescript mismatch the type definition because Angular requires `^5.0.1` but nativescript-angular restric to `~5.0.1`. This fix is to keep consistent the Rxjs version between the two repositories.